misc/HtmlTestSuite/results/test-1013.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml">
<head>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
<script src="./resources/common.js"/>
<script>
promise_test(() =&gt; {
  return waitUntilLoadedAndAutofocused().then(() =&gt; {
        assert_equals(document.activeElement, document.getElementById("outer-button"));

        var dialog = document.getElementById('dialog');
        dialog.showModal();

        autofocusButton = document.getElementById('autofocus-button');
        assert_equals(document.activeElement, autofocusButton);

        anotherButton = document.getElementById('another-button');
        anotherButton.focus();
        assert_equals(document.activeElement, anotherButton);

        // Test that recreating layout does not give focus back to a previously autofocused element.
        autofocusButton.style.display = 'none';
        document.body.offsetHeight;
        autofocusButton.style.display = 'block';
        document.body.offsetHeight;
        assert_equals(document.activeElement, anotherButton);

        // Test that reinserting does not give focus back to a previously autofocused element.
        var parentNode = autofocusButton.parentNode;
        parentNode.removeChild(autofocusButton);
        document.body.offsetHeight;
        parentNode.appendChild(autofocusButton);
        document.body.offsetHeight;
        assert_equals(document.activeElement, anotherButton);

        dialog.close();
        // Test that dialog focusing steps run when a dialog is reopened.
        dialog.showModal();
        assert_equals(document.activeElement, autofocusButton);
        dialog.close();
  });
}, "autofocus when a modal dialog is opened");
</script>
</head>
<body>
<button autofocus="" id="outer-button"/>
<dialog id="dialog">
    <button/>
    <!-- Unfocusable elements with [autofocus] should be ignored. -->
    <input autofocus="" disabled=""/>
    <textarea autofocus="" hidden=""/>
    <dialog>
        <button autofocus=""/>
    </dialog>
    <div>
        <span>
            <button autofocus="" id="autofocus-button"/>
        </span>
    </div>
    <button autofocus="" id="another-button"/>
</dialog>

</body>
</html>
